Big refactor of deep_q_networks.py
* Break down networks to components and allocate to individual file
* Add deep_q_networks_components/abc.py for refactored abstract basic classes
* Add class NetworkMixin for network structure components
* Add networks.py for network structure components
* Add loss_functions.py for loss function components
* Add optimizers.py for optimizer components
* Rename DQN classes for better consistency
