+++ evaluation.py
@@ -6,7 +6,7 @@ import numpy as np
 # Must be put before any tensorflow import statement.
 os.environ['TF_CPP_MIN_LOG_LEVEL'] = '2'
 
-from losing_connect_four.deep_q_networks import SimpleFCSgdDqn
+from losing_connect_four.deep_q_networks import Simple512SgdDqn
 from losing_connect_four.player import DeepQPlayer, Player, RandomPlayer
 from losing_connect_four.training import (
     train_one_episode, load_model_to_players,
@@ -41,7 +41,7 @@ print("\rConnect-Four Gym Environment Made")
 """Setup Players"""
 # with tf.device('/CPU:0'):
 # Setup players.
-player1: Player = DeepQPlayer(env, PARAMS, SimpleFCSgdDqn(momentum=0),
+player1: Player = DeepQPlayer(env, PARAMS, Simple512SgdDqn(momentum=0),
                               is_eval=True)
 player2: Player = RandomPlayer(env, seed=2119)
 players = {1: player1, 2: player2,
