+++ losing_connect_four/deep_q_networks.py
@@ -1,140 +1,40 @@
-from abc import ABC, abstractmethod
+"""
+Integrate DQN components to build DQN here.
+DQN built here will then be used by deep_q_model as dqn_template.
+"""
 
-from tensorflow.keras.layers import (
-    Flatten, Dense, ZeroPadding2D, Conv2D,
-    Activation, Dropout,
+from .deep_q_network_components.abc import DeepQNetwork
+from .deep_q_network_components.loss_functions import LossFuncMixinMse
+from .deep_q_network_components.networks import (
+    Simple512Net, SimpleDefaultNet, CnnShrinkNet, CnnNoShrinkNet,
 )
-from tensorflow.keras.losses import mean_squared_error
-from tensorflow.keras.models import Sequential
-from tensorflow.keras.optimizers import Adam, RMSprop, SGD
+from .deep_q_network_components.optimizers import (
+    OptimizerMixinAdam, OptimizerMixinRMSProp, OptimizerMixinSGD,
+)
 
 
-# Abstract Basic Classes.
-class DeepQNetwork(ABC):
-    """
-    Network Abstract Base Class / Mixin Class.
-
-    * create_network(...) returns the compiled network.
-    * _create_network(...) returns the network architecture.
-    * create_optimizer(...) returns the optimizer to be used.
-    * create_create_loss_function(...) returns the loss function to be used.
-    """
-    def __init__(self, *args, **kwargs):
-        self.args = args
-        self.kwargs = kwargs
-
-    def create_network(self, observation_space, action_space, params):
-        args = self.args
-        kwargs = self.kwargs
-
-        # Add channel dimension.
-        observation_space = (*observation_space, 1)
-
-        net = self._create_network(
-            observation_space, action_space, params, *args, **kwargs)
-
-        # Used Adam optimizer to allow for weight decay
-        loss_function = self.create_loss_function()
-        optimizer = self.create_optimizer(params)
-
-        net.compile(loss=loss_function, optimizer=optimizer)
-        return net
-
-    def create_optimizer(self, params):
-        args = self.args
-        kwargs = self.kwargs
-
-        return self._create_optimizer(params, *args, **kwargs)
-
-    def create_loss_function(self):
-        args = self.args
-        kwargs = self.kwargs
-
-        return self._create_loss_function(*args, **kwargs)
-
-    @abstractmethod
-    def _create_network(self, observation_space, action_space, params,
-                        *args, **kwargs):
-        raise NotImplementedError
-
-    @abstractmethod
-    def _create_optimizer(self, params, *args, **kwargs):
-        raise NotImplementedError
-
-    @abstractmethod
-    def _create_loss_function(self, *args, **kwargs):
-        raise NotImplementedError
-
-
-class OptimizerMixin(ABC):
-    """"Optimizer ABC"""
-    @abstractmethod
-    def _create_optimizer(self, params, *args, **kwargs):
-        raise NotImplementedError
-
-
-class LossFunctionMixin(ABC):
-    """"Loss Function ABC"""
-    @abstractmethod
-    def _create_loss_function(self, *args, **kwargs):
-        raise NotImplementedError
-
-
-# Optimizers.
-class OptimizerMixinAdam(OptimizerMixin):
-    def _create_optimizer(self, params, *args, **kwargs):
-        return Adam(learning_rate=params["LR"], *args, **kwargs)
-
-
-class OptimizerMixinRMSProp(OptimizerMixin):
-    def _create_optimizer(self, params, *args, **kwargs):
-        return RMSprop(learning_rate=params["LR"], *args, **kwargs)
-
-
-class OptimizerMixinSGD(OptimizerMixin):
-    def _create_optimizer(self, params, *args, **kwargs):
-        return SGD(learning_rate=params["LR"], *args, **kwargs)
-
-
-# Loss Functions.
-class LossFuncMSEMixin(LossFunctionMixin):
-    def _create_loss_function(self, *args, **kwargs):
-        return mean_squared_error
-
-
-# Networks.
-class SimpleDeepFCQNetwork(OptimizerMixinAdam, LossFuncMSEMixin, DeepQNetwork):
+class SimpleDefaultAdamDqn(
+    SimpleDefaultNet, OptimizerMixinAdam, LossFuncMixinMse, DeepQNetwork):
     """
     Architecture:
 
     - Flatten(input: observation_space)
-    - 5 Dense(2 * obs_space)
+    - 4 Dense(2 * obs_space)
     - Dense(output: action_space)
 
     Optimizer: Adam;
     Loss Function: MSE.
     """
-    def _create_network(self, observation_space, action_space, params,
-                        *args, **kwargs):
-        # Compute number of slots in Connect-Four.
-        obs_space_card = observation_space[0] * observation_space[1]
-
-        net = Sequential()
-        net.add(Flatten(input_shape=observation_space))
-        net.add(Dense(obs_space_card * 2, activation="relu"))
-        net.add(Dense(obs_space_card * 2, activation="relu"))
-        net.add(Dense(obs_space_card * 2, activation="relu"))
-        net.add(Dense(obs_space_card * 2, activation="relu"))
-        net.add(Dense(action_space, activation="linear"))
-        return net
+    pass
 
 
-class SimpleFCRMSPropDQN(OptimizerMixinRMSProp, SimpleDeepFCQNetwork):
+class Simple512RmsPropDqn(
+    Simple512Net, OptimizerMixinRMSProp, LossFuncMixinMse, DeepQNetwork):
     """
     Architecture:
 
     - Flatten(input: observation_space)
-    - 5 Dense(2 * obs_space)
+    - 4 Dense(512)
     - Dense(output: action_space)
 
     Optimizer: RMSProp;
@@ -143,12 +43,13 @@ class SimpleFCRMSPropDQN(OptimizerMixinRMSProp, SimpleDeepFCQNetwork):
     pass
 
 
-class SimpleFCSgdDqn(OptimizerMixinSGD, SimpleDeepFCQNetwork):
+class Simple512SgdDqn(
+    Simple512Net, OptimizerMixinSGD, LossFuncMixinMse, DeepQNetwork):
     """
     Architecture:
 
     - Flatten(input: observation_space)
-    - 5 Dense(2 * obs_space)
+    - 4 Dense(512)
     - Dense(output: action_space)
 
     Optimizer: SGD;
@@ -157,7 +58,8 @@ class SimpleFCSgdDqn(OptimizerMixinSGD, SimpleDeepFCQNetwork):
     pass
 
 
-class CnnDqn(LossFuncMSEMixin, OptimizerMixinSGD, DeepQNetwork):
+class CnnShrinkSgdDqn(
+    CnnShrinkNet, OptimizerMixinSGD, LossFuncMixinMse, DeepQNetwork):
     """
     Architecture:
 
@@ -179,26 +81,11 @@ class CnnDqn(LossFuncMSEMixin, OptimizerMixinSGD, DeepQNetwork):
     Optimizer: SGD;
     Loss Function: MSE.
     """
-    def _create_network(self, observation_space, action_space, params,
-                        *args, **kwargs):
-        net = Sequential()
-        # Zero-pad the top row: 6x7x1 -> 7x7x1.
-        net.add(ZeroPadding2D(padding=((1, 0), (0, 0)),
-                              input_shape=observation_space))
-        net.add(Conv2D(64, kernel_size=2, strides=1, padding="valid"))
-        net.add(Activation("relu"))  # -> 6x6x64
-        net.add(Flatten())  # -> 2304
-        net.add(Dropout(0.25))  # -> 1728
-        net.add(Dense(512, activation="relu"))  # -> 512
-        net.add(Dense(256, activation="relu"))  # -> 256
-        net.add(Dense(128, activation="relu"))  # -> 128
-        net.add(Dense(64, activation="relu"))  # -> 64
-        net.add(Dense(action_space, activation="softmax"))  # -> 7
-
-        return net
+    pass
 
 
-class CnnDqnNoShrinkage(LossFuncMSEMixin, OptimizerMixinSGD, DeepQNetwork):
+class CnnNoShrinkageSgdDqn(
+    CnnNoShrinkNet, OptimizerMixinSGD, LossFuncMixinMse, DeepQNetwork):
     """
     Architecture:
 
@@ -220,21 +107,3 @@ class CnnDqnNoShrinkage(LossFuncMSEMixin, OptimizerMixinSGD, DeepQNetwork):
     Optimizer: SGD;
     Loss Function: MSE.
     """
-
-    def _create_network(self, observation_space, action_space, params,
-                        *args, **kwargs):
-        net = Sequential()
-        # Zero-pad the top row: 6x7x1 -> 7x7x1.
-        net.add(ZeroPadding2D(padding=((1, 0), (0, 0)),
-                              input_shape=observation_space))
-        net.add(Conv2D(64, kernel_size=2, strides=1, padding="valid"))
-        net.add(Activation("relu"))  # -> 6x6x64
-        net.add(Flatten())  # -> 2304
-        net.add(Dropout(0.25))  # -> 1728
-        net.add(Dense(512, activation="relu"))  # -> 512
-        net.add(Dense(512, activation="relu"))  # -> 512
-        net.add(Dense(512, activation="relu"))  # -> 512
-        net.add(Dense(512, activation="relu"))  # -> 512
-        net.add(Dense(action_space, activation="softmax"))  # -> 7
-
-        return net
